added integration test

+++ test/Altinn.Notifications.IntegrationTests/Notifications.Core/SmsNotificationSummaryTests.cs
@@ -26,8 +26,11 @@ namespace Altinn.Notifications.IntegrationTests.Notifications.Core
 
         public async Task DisposeAsync()
         {
-            string deleteSql = $@"DELETE from notifications.orders o where o.alternateid in ('{string.Join("','", _orderIdsToDelete)}')";
-            await PostgreUtil.RunSql(deleteSql);
+            if (_orderIdsToDelete.Count != 0)
+            {
+                string deleteSql = $@"DELETE from notifications.orders o where o.alternateid in ('{string.Join("','", _orderIdsToDelete)}')";
+                await PostgreUtil.RunSql(deleteSql);
+            }
         }
 
         [Fact]
@@ -61,5 +64,27 @@ namespace Altinn.Notifications.IntegrationTests.Notifications.Core
                    throw new Exception("Expected a summary, but got an error");
                });
         }
+
+        [Fact]
+        public async Task GetSmsSummary_NoOrderIdMatchInDb_ReturnsNull()
+        {
+            SmsNotificationSummaryService service = (SmsNotificationSummaryService)ServiceUtil
+            .GetServices(new List<Type>() { typeof(ISmsNotificationSummaryService) })
+            .First(i => i.GetType() == typeof(SmsNotificationSummaryService));
+
+            // Act
+            Result<SmsNotificationSummary, ServiceError> result = await service.GetSmsSummary(Guid.NewGuid(), "ttd");
+
+            // Assert
+            result.Match(
+                success => throw new Exception("No success value should be returned if db returns null"),
+                actuallError =>
+                {
+                    Assert.IsType<ServiceError>(actuallError);
+                    Assert.Equal(404, actuallError.ErrorCode);
+                    return true;
+                });
+        }
+
     }
 }
